requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Django-advance-thumbnail==1.0.0